WRS-99 fix: Fixed note from skill to rate_info
