+++ config/doctrine/Skill.orm.yml
@@ -21,10 +21,6 @@ App\Entity\Skill:
         showNote:
             type: boolean
             nullable: true
-        note:
-            type: string
-            length: 255
-            nullable: true
         status:
             type: integer
             length: 20
